Changed every float variable to a double to ease floating point errors.
Crudely. Because Processing uses only floats and no doubles, this means
my program casts doubles to floats every time a Processing function is
invoked. I can eliminate unnessecary casts over time, as I find them, to
refine this approach a bit.

+++ src/geom/Point.java
@@ -1,9 +1,9 @@
 package geom;
 
 public class Point {
-	public float x, y;
+	public double x, y;
 	
-	public Point(float x, float y) {
+	public Point(double x, double y) {
 		this.x = x;
 		this.y = y;
 	}
